Andes Zephyr provided bundle
This is simply 'zephyr-V5.tgz' as provided by vendor unmodifed.

+++ samples/drivers/spi_flash/src/main.c
@@ -73,7 +73,7 @@ void main(void)
 	printf("\nTest 2: Flash write\n");
 	flash_write_protection_set(flash_dev, false);
 
-	printf("Attempting to write %u bytes\n", len);
+	printf("Attempting to write %lu bytes\n", (unsigned long)len);
 	rc = flash_write(flash_dev, FLASH_TEST_REGION_OFFSET, expected, len);
 	if (rc != 0) {
 		printf("Flash write failed! %d\n", rc);
